pyproject.toml
[tool.poetry]
name = "datingapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.23"
black = "^23.12.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.16"
ujson = "^5.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
